Created CategoryDtoRequest and using it in CategoryService, managed all DTOs in packages
