+++ src/main/java/com/bookshop/controller/CartController.java
@@ -1,9 +1,9 @@
 package com.bookshop.controller;
 
-import com.bookshop.dto.cart.CartDto;
-import com.bookshop.dto.cart.CartItemDtoResponse;
-import com.bookshop.dto.cart.CreateCartItemDto;
-import com.bookshop.dto.cart.PutCartItemDto;
+import com.bookshop.dto.cart.response.CartDto;
+import com.bookshop.dto.cart.response.CartItemDtoResponse;
+import com.bookshop.dto.cart.request.CreateCartItemDto;
+import com.bookshop.dto.cart.request.PutCartItemDto;
 import com.bookshop.model.User;
 import com.bookshop.res.Openapi;
 import com.bookshop.service.CartService;
