+++ src/main/java/com/bookshop/dto/book/request/BookSearchParameters.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.book;
+package com.bookshop.dto.book.request;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
