+++ src/main/java/com/bookshop/dto/cart/request/PutCartItemDto.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.cart;
+package com.bookshop.dto.cart.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
@@ -9,11 +9,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @EqualsAndHashCode
-public class CreateCartItemDto {
-    @NotNull(message = "bookId can't be null")
-    @Min(value = 0, message = "book id can't be negative number")
-    private Long bookId;
-
+public class PutCartItemDto {
     @NotNull(message = "quantity can't be null")
     @Min(value = 1, message = "quantity should be greater than 0")
     private int quantity;
