+++ src/main/java/com/bookshop/dto/cart/response/CartItemDto.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.order;
+package com.bookshop.dto.cart.response;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -7,8 +7,9 @@ import lombok.Setter;
 @Getter
 @Setter
 @EqualsAndHashCode
-public class OrderItemDto {
+public class CartItemDto {
     private Long id;
     private Long bookId;
+    private String bookTitle;
     private int quantity;
 }
