+++ src/main/java/com/bookshop/dto/cart/response/CartItemDtoResponse.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.category;
+package com.bookshop.dto.cart.response;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @EqualsAndHashCode
-public class CategoryDto {
+public class CartItemDtoResponse {
     private Long id;
-    private String name;
-    private String description;
+    private Long bookId;
+    private int quantity;
 }
