+++ src/main/java/com/bookshop/dto/order/request/ShippingAddressRequestDto.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.order;
+package com.bookshop.dto.order.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotNull;
