+++ src/main/java/com/bookshop/dto/user/request/UserRegistrationRoleRequestDto.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.user;
+package com.bookshop.dto.user.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.EqualsAndHashCode;
