+++ src/main/java/com/bookshop/dto/user/response/UserRegistrationRoleResponseDto.java
@@ -1,4 +1,4 @@
-package com.bookshop.dto.user;
+package com.bookshop.dto.user.response;
 
 import com.bookshop.model.Role;
 import io.swagger.v3.oas.annotations.media.Schema;
