+++ src/main/java/com/bookshop/service/BookService.java
@@ -1,9 +1,9 @@
 package com.bookshop.service;
 
-import com.bookshop.dto.book.BookDto;
-import com.bookshop.dto.book.BookDtoWithoutCategoryIds;
-import com.bookshop.dto.book.BookSearchParameters;
-import com.bookshop.dto.book.CreateBookRequestDto;
+import com.bookshop.dto.book.response.BookDto;
+import com.bookshop.dto.book.response.BookDtoWithoutCategoryIds;
+import com.bookshop.dto.book.request.BookSearchParameters;
+import com.bookshop.dto.book.request.CreateBookRequestDto;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
