+++ src/main/java/com/bookshop/service/CategoryService.java
@@ -1,6 +1,7 @@
 package com.bookshop.service;
 
-import com.bookshop.dto.category.CategoryDto;
+import com.bookshop.dto.category.request.CategoryDtoRequest;
+import com.bookshop.dto.category.response.CategoryDto;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
 
@@ -9,9 +10,9 @@ public interface CategoryService {
 
     CategoryDto getById(Long id);
 
-    CategoryDto save(CategoryDto categoryDto);
+    CategoryDto save(CategoryDtoRequest request);
 
-    CategoryDto update(Long id, CategoryDto categoryDto);
+    CategoryDto update(Long id, CategoryDtoRequest request);
 
     void deleteById(Long id);
 }
