+++ src/main/java/com/bookshop/service/impl/OrderServiceImpl.java
@@ -1,9 +1,9 @@
 package com.bookshop.service.impl;
 
-import com.bookshop.dto.order.OrderDto;
-import com.bookshop.dto.order.OrderItemDto;
-import com.bookshop.dto.order.ShippingAddressRequestDto;
-import com.bookshop.dto.order.StatusRequestDto;
+import com.bookshop.dto.order.response.OrderDto;
+import com.bookshop.dto.order.response.OrderItemDto;
+import com.bookshop.dto.order.request.ShippingAddressRequestDto;
+import com.bookshop.dto.order.request.StatusRequestDto;
 import com.bookshop.exception.EntityNotFoundException;
 import com.bookshop.mapper.OrderItemMapper;
 import com.bookshop.mapper.OrderMapper;
