+++ src/main/resources/db/changelog/changes/06-create-categories-table.yaml
@@ -18,6 +18,7 @@ databaseChangeLog:
                   type: varchar(255)
                   constraints:
                     nullable: false
+                    unique: true
               - column:
                   name: description
                   type: varchar(255)
